+++ app/src/main/java/com/dosei/games/toybank/feature/game/play/GameplayViewModel.kt
@@ -2,7 +2,6 @@ package com.dosei.games.toybank.feature.game.play
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.dosei.games.toybank.data.local.storage.player.Player
 import com.dosei.games.toybank.data.repository.GameplayRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
