+++ app/src/main/java/com/dosei/games/toybank/navigation/AppNavigation.kt
@@ -6,6 +6,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.dosei.games.toybank.AppRoutes
+import com.dosei.games.toybank.feature.game.play.GameplayScreen
 import com.dosei.games.toybank.feature.game.setup.GameSetupScreen
 import com.dosei.games.toybank.feature.home.HomeScreen
 
@@ -24,7 +25,7 @@ fun AppNavigation(controller: NavHostController) {
         }
 
         composable<AppRoutes.Gameplay> {
-
+            GameplayScreen(controller, hiltViewModel())
         }
     }
 }
